working update window
