+++ Assets/Scenes/Level1.unity
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:78f4dd015c01ef49ba6d1b8b501d1c5113ab6b475cd88a7c941850fc78ce7d06
-size 122834
+oid sha256:27d21fd7b7c94233700e5f3a9ede6c01d36df2f13e039011d23e15fe25c74fdb
+size 124891
