+++ Assets/Scripts/updateWindow.cs
@@ -1,80 +1,81 @@
-//// By Adam Nixdorf
-//using System;
-//using UnityEngine;
-//using UnityEngine.UIElements;
+// By Adam Nixdorf
+using System;
+using UnityEngine;
+using UnityEngine.UIElements;
 
-////public class updateWindow : MonoBehaviour 
-//{
-    // assign the names of the UI elements in the UI Builder
-    //public static updateWindow instance;
-    //private string updateInfoWindowName = "updateInfoWindow";
-    //private string updateVisElementName = "updateVisualElement";
-    //private string updateLabelName = "updateLabel";
-    //private string continueButtonName = "continueButton";
+public class updateWindow : MonoBehaviour
+{
+    //assign the names of the UI elements in the UI Builder
+    public static updateWindow instance;
+    private string updateInfoWindowName = "updateInfoWindow";
+    private string updateVisElementName = "updateVisElement";
+    private string updateLabelName = "updateLabel";
+    private string continueButtonName = "continueButton";
 
-    // assign variables to the UI elements
-    //private string updateLabelText;
-    //private Label updateLabel;
-    //private Button continueButton;
-    //private VisualElement updateVisElement;
-    //private VisualElement updateInfoWindow;
-
-
-    //private void Awake()
-    //{
-        //if (instance == null)
-        //{
-        //    instance = this;
-        //}
-        //else if (instance != this)
-        //{
-        //    Destroy(gameObject);
-        //}
-        
-        //updateInfoWindow = GetComponent<UIDocument>().rootVisualElement.Q<VisualElement>(updateInfoWindowName);
-        //updateVisElement = updateInfoWindow.Q<VisualElement>(updateVisElementName);
-        //updateLabel = updateVisElement.Q<Label>(updateLabelName);
-        //continueButton = updateInfoWindow.Q<Button>(continueButtonName);
-    //}
-    //internal void DisplayUpdateWindow(string displayText)
-    //{
-    //    updateInfoWindow.style.display = DisplayStyle.Flex; // Show the update info window
-
-    //    Debug.Log("Displaying update window with text: " + displayText);
-
-    //    if (updateLabel == null)
-    //    {
-    //        Debug.LogError("updateLabel is null");  
-    //    }
-    //    if (updateInfoWindow == null)
-    //    {
-    //        Debug.LogError("updateInfoWindow is null");
-    //    }
-    //    if (continueButton == null)
-    //    {
-    //        Debug.LogError("continueButton is null");
-    //    }
-    //    if (updateVisElement == null)
-    //    {
-    //        Debug.LogError("updateVisElement is null");
-    //    }
-    //    if (updateLabel == null || updateInfoWindow == null || continueButton == null)
-    //    {
-    //        return; // Exit if any of the elements are null
-    //    }
-    //    updateLabel.text = displayText;
-
-    //    continueButton.clicked += () =>
-    //    {
-    //        updateInfoWindow.style.display = DisplayStyle.None;
+    //assign variables to the UI elements
+    private string updateLabelText;
+    private Label updateLabel;
+    private Button continueButton;
+    private VisualElement updateVisElement;
+    private VisualElement updateInfoWindow;
 
 
-    //        if (PlayerInformation.control.isPaused == true)
-    //        {
-    //            PauseGame.pauseResumeGame();
-    //        }
-    //    };
-    //}
+    private void Awake()
+    {
+        if (instance == null)
+        {
+            instance = this;
+        }
+        else if (instance != this)
+        {
+            Destroy(gameObject);
+        }
+
+        updateInfoWindow = GetComponent<UIDocument>().rootVisualElement.Q<VisualElement>(updateInfoWindowName);
+        updateVisElement = updateInfoWindow.Q<VisualElement>(updateVisElementName);
+        updateLabel = updateVisElement.Q<Label>(updateLabelName);
+        continueButton = updateInfoWindow.Q<Button>(continueButtonName);
+    }
+    internal void DisplayUpdateWindow(string displayText)
+    {
+        updateInfoWindow.style.display = DisplayStyle.Flex; // Show the update info window
+
+        Debug.Log("Displaying update window with text: " + displayText);
+
+        if (updateLabel == null)
+        {
+            Debug.LogError("updateLabel is null");
+        }
+        if (updateInfoWindow == null)
+        {
+            Debug.LogError("updateInfoWindow is null");
+        }
+        if (continueButton == null)
+        {
+            Debug.LogError("continueButton is null");
+        }
+        if (updateVisElement == null)
+        {
+            Debug.LogError("updateVisElement is null");
+        }
+        if (updateLabel == null || updateInfoWindow == null || continueButton == null)
+        {
+            return; // Exit if any of the elements are null
+        }
+        updateLabel.text = displayText;
+
+        continueButton.clicked += () =>
+        {
+            updateInfoWindow.style.display = DisplayStyle.None;
+
+
+            if (PlayerInformation.control.isPaused == true)
+            {
+                PauseGame.pauseResumeGame();
+            }
+        };
+    }
+}
 
     
-//}
+
