+++ Assets/UI Toolkit/UI/UITemplate.uxml
@@ -42,12 +42,4 @@
             <ui:Button text="Continue" name="continueButton" class="button" style="align-self: auto; flex-direction: row; justify-content: flex-start;" />
         </ui:VisualElement>
     </ui:VisualElement>
-    <ui:VisualElement name="updateInfoWindow" style="flex-grow: 1; background-image: url(&quot;project://database/Assets/Fantasy%20Wooden%20GUI/normal_ui_set%20A/UI%20board%20Large%20Set.png?fileID=2800000&amp;guid=90a2dde0942123c4ebb9c4af580c5818&amp;type=3#UI board Large Set&quot;); margin-top: 5%; margin-right: 10%; margin-bottom: 10%; margin-left: 10%; padding-top: 20px; padding-right: 20px; padding-bottom: 20px; padding-left: 20px; flex-direction: column; justify-content: flex-start; align-content: center; flex-wrap: wrap; -unity-font-definition: url(&quot;project://database/Assets/TextMesh%20Pro/Examples%20&amp;%20Extras/Fonts/BMYEONSUNG_ttf.ttf?fileID=12800000&amp;guid=ee9e312b956c7df4186ade92cf957db7&amp;type=3#BMYEONSUNG_ttf&quot;); color: rgb(110, 90, 67); font-size: 16px; -unity-font-style: bold; display: none;">
-        <ui:VisualElement name="updateVisElement" style="flex-grow: 1; padding-top: 20px; padding-right: 30px; padding-bottom: 20px; padding-left: 30px; flex-direction: row; align-content: stretch; flex-wrap: wrap; justify-content: space-evenly;">
-            <ui:Label name="updatelLabel" text="test" style="flex-direction: column; align-items: center; align-content: auto;" />
-        </ui:VisualElement>
-        <ui:VisualElement name="updateWindowButtons" style="flex-grow: 1; flex-direction: row; align-self: auto; align-items: auto; padding-top: 20px; padding-right: 20px; padding-bottom: 20px; padding-left: 20px; bottom: auto; top: auto;">
-            <ui:Button text="Continue" name="continueButton" class="button" style="align-self: auto; flex-direction: row; justify-content: flex-start;" />
-        </ui:VisualElement>
-    </ui:VisualElement>
 </ui:UXML>
