+++ Assets/UI Toolkit/UpdateWindow.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7fbe6be132b79a94e808fc09fe45cb0f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
