+++ Assets/UI Toolkit/UpdateWindow/UpdateWindow.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8965a07e3ae39aa48bf549c9d0287788
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
