+++ Assets/UI Toolkit/UpdateWindow/UpdateWindow.uxml
@@ -1,0 +1,11 @@
+<ui:UXML xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:ui="UnityEngine.UIElements" xmlns:uie="UnityEditor.UIElements" noNamespaceSchemaLocation="../../../UIElementsSchema/UIElements.xsd" editor-extension-mode="False">
+    <Style src="project://database/Assets/UI%20Toolkit/HealthQuiz/HealthStyles.uss?fileID=7433441132597879392&amp;guid=dcacbd141cbff95409f2c7a11988c968&amp;type=3#HealthStyles" />
+    <ui:VisualElement name="updateInfoWindow" style="flex-grow: 1; background-image: url(&quot;project://database/Assets/Fantasy%20Wooden%20GUI/normal_ui_set%20A/UI%20board%20Large%20Set.png?fileID=2800000&amp;guid=90a2dde0942123c4ebb9c4af580c5818&amp;type=3#UI board Large Set&quot;); margin-top: 5%; margin-right: 10%; margin-bottom: 10%; margin-left: 10%; padding-top: 20px; padding-right: 20px; padding-bottom: 20px; padding-left: 20px; flex-direction: column; justify-content: flex-start; align-content: center; flex-wrap: wrap; -unity-font-definition: url(&quot;project://database/Assets/TextMesh%20Pro/Examples%20&amp;%20Extras/Fonts/BMYEONSUNG_ttf.ttf?fileID=12800000&amp;guid=ee9e312b956c7df4186ade92cf957db7&amp;type=3#BMYEONSUNG_ttf&quot;); color: rgb(110, 90, 67); font-size: 16px; -unity-font-style: bold; display: none;">
+        <ui:VisualElement name="updateVisElement" style="flex-grow: 1; padding-top: 20px; padding-right: 30px; padding-bottom: 20px; padding-left: 30px; flex-direction: row; align-content: stretch; flex-wrap: wrap; justify-content: space-evenly; margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%;">
+            <ui:Label name="updateLabel" text="test" style="flex-direction: column; align-items: center; align-content: auto;" />
+        </ui:VisualElement>
+        <ui:VisualElement name="updateWindowButtons" style="flex-grow: 1; flex-direction: row; align-self: auto; align-items: auto; padding-top: 20px; padding-right: 20px; padding-bottom: 20px; padding-left: 20px; bottom: auto; top: auto;">
+            <ui:Button text="Continue" name="continueButton" class="button" style="align-self: auto; flex-direction: row; justify-content: flex-start;" />
+        </ui:VisualElement>
+    </ui:VisualElement>
+</ui:UXML>
