+++ Assets/UI Toolkit/UpdateWindow/UpdateWindow.uxml.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 9327e275472ddb84a920a0b67a041efc
+ScriptedImporter:
+  internalIDToNameTable: []
+  externalObjects: {}
+  serializedVersion: 2
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 13804, guid: 0000000000000000e000000000000000, type: 0}
